FIX: Explicit dtypes

+++ entity_linking/data/mention_entity_matching.py
@@ -17,9 +17,9 @@ class MentionEntityMatchingDataset(Dataset):
     def __getitem__(self, idx):
         item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
         item.update({
-            'mention_spans': torch.tensor(self.mention_spans[idx], dtype=torch.int),
-            'entity_indices': torch.tensor(self.entity_indices[idx], dtype=torch.int),
-            'labels': torch.tensor(self.entity_labels[idx], dtype=torch.int)
+            'mention_spans': torch.tensor(self.mention_spans[idx]),
+            'entity_indices': torch.tensor(self.entity_indices[idx]),
+            'labels': torch.tensor(self.entity_labels[idx])
         })
         return item
 
@@ -83,9 +83,9 @@ def _filter_truncated_entities(tokens: List[List[str]], entity_info: List[List[T
 
 def _process_entity_info(entity_info: List[List[Tuple[int, Tuple[int, int], list]]], offset_mapping: List[List[Tuple[int, int]]], num_ents: int, items_per_chunk: int) -> tuple:
     num_chunks = len(entity_info)
-    mention_spans = np.zeros((num_chunks, num_ents, 2))
-    entity_indices = np.zeros((num_chunks, num_ents))
-    entity_labels = np.ones((num_chunks, num_ents)) * -100  # initialize all with label: ignore
+    mention_spans = np.zeros((num_chunks, num_ents, 2), dtype=int)
+    entity_indices = np.zeros((num_chunks, num_ents), dtype=int)
+    entity_labels = np.ones((num_chunks, num_ents), dtype=int) * -100  # initialize all with label: ignore
     ents_per_item = num_ents // items_per_chunk
     # prepare word blocker to retrieve entities with similar surface forms
     dbr = DbpResourceStore.instance()
